🚚(frontend) move ContentCard to lib-components
As we want to use ContentCard in LTI select, we have to extract it from
the standalone_site.
