+++ src/frontend/apps/standalone_site/src/components/Modal/Modal.tsx
@@ -2,6 +2,7 @@ import { Box, Layer, LayerProps } from 'grommet';
 import { normalizeColor } from 'grommet/utils';
 import { FormClose } from 'grommet-icons';
 import { Nullable, theme } from 'lib-common';
+import { useResponsive } from 'lib-components';
 import {
   forwardRef,
   MutableRefObject,
@@ -12,8 +13,6 @@ import {
 } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 import styled from 'styled-components';
-
-import { useResponsive } from 'hooks/useResponsive';
 
 const messages = defineMessages({
   closeButtonTitle: {
