+++ src/frontend/apps/standalone_site/src/features/Contents/features/ClassRoom/components/Create/ClassRoomCreate.tsx
@@ -1,11 +1,11 @@
 import { Button, Heading, Text } from 'grommet';
+import { useResponsive } from 'lib-components';
 import { Fragment } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 import { Link, Route, Switch, useHistory } from 'react-router-dom';
 
 import { WhiteCard } from 'components/Cards';
 import { Modal } from 'components/Modal';
-import { useResponsive } from 'hooks/useResponsive';
 import { routes } from 'routes';
 
 import ClassroomCreateForm from './ClassRoomCreateForm';
