+++ src/frontend/apps/standalone_site/src/features/Contents/index.tsx
@@ -1,4 +1,3 @@
-export * from './components/ContentCard/ContentCard';
 export { default as useContentPerPage } from './hooks/useContentPerPage';
 export { default as ContentsWrapper } from './components/ContentsWrapper/ContentsWrapper';
 export { default as ContentsRouter } from './components/ContentsRouter/ContentsRouter';
