+++ src/frontend/apps/standalone_site/src/features/Menu/components/Burger/Burger.tsx
@@ -1,8 +1,8 @@
+import { useResponsive } from 'lib-components';
 import { SVGProps } from 'react';
 
 import { ReactComponent as BurgerIcon } from 'assets/svg/burger.svg';
 import { useMenu } from 'features/Menu/store/menuStore';
-import { useResponsive } from 'hooks/useResponsive';
 
 import styles from './Burger.module.css';
