+++ src/frontend/apps/standalone_site/src/styles/theme.extend.ts
@@ -1,5 +1,5 @@
 import { deepMerge } from 'grommet/utils';
-import { theme } from 'lib-common';
+import { theme, breakpoints } from 'lib-common';
 import { css } from 'styled-components';
 
 export const themeExtend = {
@@ -20,13 +20,7 @@ export const themeExtend = {
         even: '0px 0px 6px 0px rgba(2, 117, 180, 0.3)',
       },
     },
-    breakpoints: {
-      xxsmall: { value: 380 },
-      xsmall: { value: 540 },
-      xsmedium: { value: 1024 },
-      smedium: { value: 1280 },
-      large: { value: 9999 },
-    },
+    breakpoints: breakpoints,
   },
 };
