+++ src/frontend/packages/lib_components/src/common/ContentCard/index.spec.tsx
@@ -1,6 +1,7 @@
 import { render, screen } from '@testing-library/react';
+import React from 'react';
 
-import { ContentCard } from './ContentCard';
+import { ContentCard } from '.';
 
 describe('<ContentCard />', () => {
   test('renders ContentCard', () => {
