+++ src/frontend/packages/lib_components/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './stores';
+export * from './useResponsive';
